Do not recurse into replacement nested catalogs when merging
When replacing an existing nested catalog with a new nested catalog,
there is no need for the merge operation to recurse into both old and
new nested catalogs and perform updates on an entry-by-entry basis.
The merge may be effected simply by updating the appropriate entry
within the nested_catalogs database table (and by updating the
statistics counters for the parent catalog).

When using a gateway to handle large volumes of grafted files and with
the lease paths aligned with the nested catalog mountpoints, this
optimisation can reduce the amount of work performed by the gateway to
trivial levels: instead of handling a few hundred thousand individual
directory entry catalog merge operations, the gateway needs to perform
only a single nested catalog replacement.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>