Check for writable mountpoint in packaging scriptlets
The packaging scriptlets currently check for the existence of any
version of the CVMFS transaction.lock file or subdirectory.  Check
instead for a writable mountpoint, in preparation for eliminating the
transaction.lock file.

This check remains correct when upgrading from older packaged versions
of CVMFS, since the mountpoint will always be writable when a
transaction is in progress.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>