ABANDON - Eliminate the transaction.lock file
The transaction.lock file is a flag file (not a lock file) that
indicates whether or not a transaction is currently in progress.

The same information is already present in the read-write status of
the mountpoint: if the mountpoint is writable then a transaction is in
progress.

Simplify the code by eliminating the transaction.lock file and instead
checking the read-write status of the mountpoint as needed.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>