Set CVMFS_ROOT_HASH for the duration of a transaction
A transaction represents a set of operations carried out while the
union mountpoint is writable, during which modifications may be
accumulated in the overlay filesystem.  It is necessary that the
underlying read-only mountpoint remains unchanged for the duration of
the transaction, so that the accumulated changes represent changes
from a known catalog revision.

The current logic sets CVMFS_ROOT_HASH at all times unless a gateway
is being used, in which case CVMFS_ROOT_HASH will never be set.  This
creates a race condition when using a gateway, since the underlying
read-only mountpoint may change state mid-transaction.  It is also
redundant when not using a gateway and outside a transaction (and not
using a checkout marker), since in those circumstances the published
root hash cannot change anyway.

Redefine the use of CVMFS_ROOT_HASH in client.local as follows:

- when inside a transaction, CVMFS_ROOT_HASH is fixed to represent the
  catalog at the time that the transaction was opened

- when outside a transaction, CVMFS_ROOT_HASH is left undefined unless
  a checkout marker is present, in which case it will have the value
  from the checkout marker.

This may be efficiently implemented using the algorithm:

- upon opening a transaction, set CVMFS_ROOT_HASH to the value of the
  currently mounted catalog

- upon closing a transaction, clear CVMFS_ROOT_HASH unless a checkout
  marker is still present, then unmount and remount the repository

- to perform a checkout: open a transaction, create the checkout
  marker, set CVMFS_ROOT_HASH, and close the transaction (which will
  implicitly remount the repository at the new CVMFS_ROOT_HASH value)

- to undo a checkout: open a transaction, remove the checkout marker,
  and close the transaction (which will implicitly remount the
  repository at the latest published revision)

This results in almost all of the relevant logic being encapsulated
within the open_transaction() and close_transaction() functions, with
only cvmfs_server_checkout() having to worry about direct manipulation
of the root hash.

Update the code to reflect this simpler (and less race-prone) usage
pattern.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>